doc no longer depend on clap
