+++ lang/docs/Cargo.toml
@@ -10,7 +10,6 @@ repository.workspace = true
 categories.workspace = true
 
 [dependencies]
-clap = { version = "4", features = ["derive"] }
 # fancy error messages
 miette = { workspace = true, features = ["fancy"] }
 opener = "0.5"
