+++ lang/docs/src/lib.rs
@@ -1,1 +1,4 @@
 pub mod doc;
+
+pub use doc::open;
+pub use doc::write_html;
